+++ reper_front/app/src/main/java/com/ssafy/reper/data/dto/LoginResponse.kt
@@ -1,8 +1,8 @@
 package com.ssafy.reper.data.dto
 
-data class LoginResponse (
-    val role : String? = null,
+data class LoginResponse(
+    val role: String? = null,
     val loginIdCookie: String? = null,
-    val userId: Long? = null,
+    val userId: Int? = null,
     val username: String? = null
 )
